Resolved the image sizing issue, must now figure out Tensors for Conv2D
There was an issue with the padding implementation for the resizing that needed to be re adjusted, and so that has been resolved.

What now needs to be resolved next is the tensor input into the Conv2D model as the model takes in a single Tensor, but the image files are read in as 80 tensors. Will resolve later.